LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/window-scroll-promises-expected.txt


FAIL Window.scroll(auto): return type assert_true: expected true got false
PASS Window.scroll(auto): promise behavior
FAIL Window.scroll(instant): return type assert_true: expected true got false
PASS Window.scroll(instant): promise behavior
FAIL Window.scroll(smooth): return type assert_true: expected true got false
FAIL Window.scroll(smooth): promise behavior assert_equals: Position after await expected 5000 but got 0
FAIL Window.scrollTo(auto): return type assert_true: expected true got false
PASS Window.scrollTo(auto): promise behavior
FAIL Window.scrollTo(instant): return type assert_true: expected true got false
PASS Window.scrollTo(instant): promise behavior
FAIL Window.scrollTo(smooth): return type assert_true: expected true got false
FAIL Window.scrollTo(smooth): promise behavior assert_equals: Position after await expected 5000 but got 0
FAIL Window.scrollBy(auto): return type assert_true: expected true got false
PASS Window.scrollBy(auto): promise behavior
FAIL Window.scrollBy(instant): return type assert_true: expected true got false
PASS Window.scrollBy(instant): promise behavior
FAIL Window.scrollBy(smooth): return type assert_true: expected true got false
FAIL Window.scrollBy(smooth): promise behavior assert_equals: Position after await expected 5000 but got 0
